Implemented pausing system. So that it is controlled by gavel connect not easylive.
added trim to getLot function to ensure that blank spaces don't interfere with logic.

Auction platform will now add the current lot easylive is on. Provided if it is not already in the system. I would like to change this in the future, as I don't see how skipping lots is going to be done.

System will automatically close the auction start window.